Prefix types in namespaces other than pg_catalog or public
It's possible to define, e.g., an enum with the same name in two
different schemas. When initializing data types, types defined within
schemas other than pg_catalog or public should be prefixed by the schema
name and a dot to disambiguate them and ensure all types with the same
base name get added to the map of OID to type. Prior to this commit, the
last type scanned would "win", and all others would be missing from the
ConnInfo type maps, which would subsequently cause and PREPARE involving
columns of the missing types to return the error "unknown oid".